Former-commit-id: ea74e578082314970f2538a977e7f4fb73610eba

+++ src/gfx/Cuboid.cpp
@@ -44,9 +44,9 @@ void Pointer<Cuboid::Vertex>()
 Cuboid::Cuboid()
 {
 	Vertex verts[3][2][4];
-	GLuint indices[3][2][2][3];
+	GLushort indices[3][2][2][3];
 	
-	GLuint * index = * * * indices;
+	GLushort * index = * * * indices;
 	for (int axis = 0; axis < 3; ++ axis)
 	{
 		for (int pole = 0; pole < 2; ++ pole)
@@ -73,7 +73,7 @@ Cuboid::Cuboid()
 				}
 			}
 			
-			GLuint index_base = polygon_verts - verts[0][0];
+			GLushort index_base = polygon_verts - verts[0][0];
 			* (index ++) = index_base + 0;
 			* (index ++) = index_base + 1;
 			* (index ++) = index_base + 2;
@@ -85,7 +85,7 @@ Cuboid::Cuboid()
 	}
 	
 	int num_verts = sizeof(verts) / sizeof(Vertex);
-	int num_indices = sizeof(indices) / sizeof(GLuint);
+	int num_indices = sizeof(indices) / sizeof(GLushort);
 	
 	_mesh.Init();
 	_mesh.Bind();
