+++ src/gfx/MeshBufferObject.h
@@ -60,7 +60,7 @@ namespace gfx
 			vbo.BufferSubData(num, array);
 		}
 		
-		void SetIbo(int num, GLuint const * array)
+		void SetIbo(int num, GLushort const * array)
 		{
 			ibo.BufferSubData(num, array);
 		}
